+++ packages/scaffold-ui/src/partials/w3m-connecting-wc-qrcode/index.ts
@@ -72,14 +72,12 @@ export class W3mConnectingWcQrcode extends W3mConnectingWidget {
     ConnectionController.setWcLinking(undefined)
     ConnectionController.setRecentWallet(this.wallet)
 
-    const themeVariables = ThemeController.state.themeVariables
-
     return html` <wui-qr-code
       size=${size}
       theme=${ThemeController.state.themeMode}
       uri=${this.uri}
       imageSrc=${ifDefined(AssetUtil.getWalletImage(this.wallet))}
-      color=${ifDefined(themeVariables['--w3m-qr-color'])}
+      color=${ifDefined(ThemeController.state.themeVariables['--w3m-qr-color'])}
       alt=${ifDefined(alt)}
       data-testid="wui-qr-code"
     ></wui-qr-code>`
