--- packages/wallet-button/readme.md
@@ -1,14 +1,0 @@
-#### 📚 [Documentation](https://docs.reown.com/2.0/appkit/about)
-
-#### 🔗 [Website](https://reown.com/appkit)
-
-# AppKit SIWX
-
-The chain agnostic tool to enable authentication for AppKit applications.
-
-> [!WARNING]
-> This package is still in development and should not be used in production.
-
-<p align="center">
-  <img src="./.github/assets/header.png" alt="" border="0">
-</p>
